move fixtures to e2e folder
