+++ e2e/specs/skip-onboarding.spec.js
@@ -1,21 +1,26 @@
 'use strict';
-import { Regression } from '../tags';
+import { Smoke } from '../tags';
 
 import WalletView from '../pages/WalletView';
 
 import LoginView from '../pages/LoginView';
+import { withFixtures } from '../fixtures/fixture-helper';
+import FixtureBuilder from '../fixtures/fixture-builder';
 
 const PASSWORD = '123123123';
 
-describe(Regression('Skip Onboarding to test fixtures'), () => {
-  beforeEach(() => {
+describe(Smoke('Skip Onboarding to test fixtures'), () => {
+  beforeAll(async () => {
     jest.setTimeout(150000);
   });
 
   it('should launch the app and log in', async () => {
-    await LoginView.isVisible();
-    await LoginView.enterPassword(PASSWORD);
+    const fixture = new FixtureBuilder().build();
+    await withFixtures({ fixture, restartDevice: true }, async () => {
+      await LoginView.isVisible();
+      await LoginView.enterPassword(PASSWORD);
 
-    await WalletView.isVisible();
+      await WalletView.isVisible();
+    });
   });
 });
