+++ e2e/viewHelper.js
@@ -22,12 +22,15 @@ import TestHelpers from './helpers';
 
 import TermsOfUseModal from './pages/modals/TermsOfUseModal';
 import TabBarComponent from './pages/TabBarComponent';
+import LoginView from './pages/LoginView';
+import { withFixtures } from './fixtures/fixture-helper';
+import FixtureBuilder from './fixtures/fixture-builder';
 
 const GOERLI = 'Goerli Test Network';
 
 const LOCALHOST_URL = 'http://localhost:8545/';
 
-// detox on ios does not have a clean way of interacting with webview eleemnts. You would need to tap by coordinates
+// detox on ios does not have a clean way of interacting with webview elements. You would need to tap by coordinates
 export const testDappConnectButtonCooridinates = { x: 170, y: 280 };
 export const testDappSendEIP1559ButtonCoordinates = { x: 320, y: 500 };
 const validAccount = Accounts.getValidAccount();
@@ -180,4 +183,15 @@ export const switchToGoreliNetwork = async () => {
   await NetworkListModal.changeNetwork(GOERLI);
   await WalletView.isNetworkNameVisible(GOERLI);
   await NetworkEducationModal.tapGotItButton();
+};
+
+export const skipOnboarding = async () => {
+  // const fixture = new FixtureBuilder().build();
+  // await withFixtures({ fixture, restartDevice: true }, async () => {
+    const PASSWORD = '123123123';
+    await LoginView.isVisible();
+    await LoginView.enterPassword(PASSWORD);
+
+    await WalletView.isVisible();
+  // });
 };
