--- wdio/fixtures/fixture-builder.js
@@ -1,528 +1,0 @@
-/**
- * FixtureBuilder class provides a fluent interface for building fixture data.
- */
-class FixtureBuilder {
-    /**
-    * Create a new instance of FixtureBuilder.
-    * @param {Object} options - Options for the fixture builder.
-    * @param {boolean} options.onboarding - Flag indicating if onboarding fixture should be used.
-    */
-     constructor({onboarding = false} = {}) {
-    // Initialize the fixture based on the onboarding flag
-      onboarding === true ? this.withOnboardingFixture() : this.withDefaultFixture();
-    }
-  
-    /**
-     * Set the asyncState property of the fixture.
-     * @param {any} asyncState - The value to set for asyncState.
-     * @returns {FixtureBuilder} - The FixtureBuilder instance for method chaining.
-     */
-    withAsyncState(asyncState) {
-      this.fixture.asyncState = asyncState;
-      return this;
-    }
-  
-    /**
-     * Set the state property of the fixture.
-     * @param {any} state - The value to set for state.
-     * @returns {FixtureBuilder} - The FixtureBuilder instance for method chaining.
-     */
-    withState(state) {
-      this.fixture.state = state;
-      return this;
-    }
-  
-    /**
-     * Set the default fixture values.
-     * @returns {FixtureBuilder} - The FixtureBuilder instance for method chaining.
-     */
-    withDefaultFixture() {
-      this.fixture = {
-            state: {
-              collectibles: {
-                favorites: {}
-              },
-              engine: {
-                backgroundState: {
-                  AccountTrackerController: {
-                    accounts: {
-                      '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3': {
-                        balance: "0x0"
-                      }
-                    },
-                    _U: 0,
-                    _V: 1,
-                    _X: null
-                  },
-                  AddressBookController: {
-                    addressBook: {}
-                  },
-                  AssetsContractController: {},
-                  NftController: {
-                    allNftContracts: {},
-                    allNfts: {},
-                    ignoredNfts: []
-                  },
-                  TokenListController: {
-                    tokenList: {
-                      '0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f': {
-                        address: "0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f",
-                        symbol: "SNX",
-                        decimals: 18,
-                        name: "Synthetix Network Token",
-                        iconUrl: "https://static.metafi.codefi.network/api/v1/tokenIcons/1/0xc011a73ee8576fb46f5e1c5751ca3b9fe0af2a6f.png",
-                        type: "erc20",
-                        aggregators: [
-                          "Aave",
-                          "Bancor",
-                          "CMC",
-                          "Crypto.com",
-                          "CoinGecko",
-                          "1inch",
-                          "PMM",
-                          "Synthetix",
-                          "Zerion",
-                          "Lifi"
-                        ],
-                        occurrences: 10,
-                        fees: {
-                          '0x5fd79d46eba7f351fe49bff9e87cdea6c821ef9f': 0,
-                          '0xda4ef8520b1a57d7d63f1e249606d1a459698876': 0
-                        }
-                      }
-                    },
-                    tokensChainsCache: {
-                      1: {}
-                    },
-                    preventPollingOnNetworkRestart: false
-                  },
-                  CurrencyRateController: {
-                    conversionDate: 1684232383.997,
-                    conversionRate: 1815.41,
-                    nativeCurrency: "ETH",
-                    currentCurrency: "usd",
-                    pendingCurrentCurrency: null,
-                    pendingNativeCurrency: null,
-                    usdConversionRate: 1815.41
-                  },
-                  KeyringController: {
-                    vault: "{\"cipher\":\"ynNI8tAH4fcpmXo8S88A/3T3Dd1w0LY5ftpL59gW0ObYxovgFhrtKpRe/WD7WU42KwGBNKVicB9W9at4ePgOJGS6IMWr//C3jh0vKQTabkDzDy1ZfSvztRxGpVjmrnU3fC5B0eq/MBMSrgu8Bww309pk5jghyRfzp9YsG0ONo1CXUm2brQo/eRve7i9aDbiGXiEK0ch0BO7AvZPGMhHtYRrrOro4QrDVHGUgAF5SA1LD4dv/2AB8ctHwn4YbUmICieqlhJhprx3CNOJ086g7vPQOr21T4IbvtTumFaTibfoD3GWHQo11CvE04z3cN3rRERriP7bww/tZOe8OAMFGWANkmOJHwPPwEo1NBr6w3GD2VObEmqNhXeNc6rrM23Vm1JU40Hl+lVKubnbT1vujdGLmOpDY0GdekscQQrETEQJfhKlXIT0wwyPoLwR+Ja+GjyOhBr0nfWVoVoVrcTUwAk5pStBMt+5OwDRpP29L1+BL9eMwDgKpjVXRTh4MGagKYmFc6eKDf6jV0Yt9pG+jevv5IuyhwX0TRtfQCGgRTtS7oxhDQPxGqu01rr+aI7vGMfRQpaKEEXEWVmMaqCmktyUV35evK9h/xv1Yif00XBll55ShxN8t2/PnATvZxFKQfjJe5f/monbwf8rpfXHuFoh8M9hzjbcS5eh/TPYZZu1KltpeHSIAh5C+4aFyZw0e1DeAg/wdRO3PhBrVztsHSyISHlRdfEyw7QF4Lemr++2MVR1dTxS2I5mUEHjh+hmp64euH1Vb/RUppXlmE8t1RYYXfcsF2DlRwPswP739E/EpVtY3Syf/zOTyHyrOJBldzw22sauIzt8Q5Fe5qA/hGRWiejjK31P/P5j7wEKY7vrOJB1LWNXHSuSjffx9Ai9E\",\"iv\":\"d5dc0252424ac0c08ca49ef320d09569\",\"salt\":\"feAPSGdL4R2MVj2urJFl4A==\",\"lib\":\"original\"}",
-                    keyrings: [
-                      {
-                        accounts: [
-                          "0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3"
-                        ],
-                        index: 0,
-                        type: "HD Key Tree"
-                      },
-                      {
-                        accounts: [],
-                        index: 1,
-                        type: "QR Hardware Wallet Device"
-                      }
-                    ]
-                  },
-                  PersonalMessageManager: {
-                    unapprovedMessages: {},
-                    unapprovedMessagesCount: 0
-                  },
-                  NetworkController: {
-                    network: "1",
-                    isCustomNetwork: false,
-                    providerConfig: {
-                      type: "mainnet",
-                      chainId: "1"
-                    },
-                    networkDetails: {
-                      isEIP1559Compatible: true
-                    }
-                  },
-                  PhishingController: {
-                    listState: {
-                      allowlist: [],
-                      fuzzylist: [
-                        "cryptokitties.co",
-                        "launchpad.ethereum.org",
-                        "etherscan.io",
-                        "makerfoundation.com",
-                        "metamask.io",
-                        "myetherwallet.com",
-                        "opensea.io",
-                        "satoshilabs.com"
-                      ],
-                      version: 2,
-                      name: "MetaMask",
-                      tolerance: 1,
-                      lastUpdated: 1684231917
-                    },
-                    whitelist: [],
-                    hotlistLastFetched: 1684231917,
-                    stalelistLastFetched: 1684231917
-                  },
-                  PreferencesController: {
-                    featureFlags: {},
-                    frequentRpcList: [],
-                    identities: {
-                      '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3': {
-                        address: "0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3",
-                        name: "Account 1",
-                        importTime: 1684232000456
-                      }
-                    },
-                    ipfsGateway: "https://cloudflare-ipfs.com/ipfs/",
-                    lostIdentities: {},
-                    selectedAddress: "0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3",
-                    useTokenDetection: true,
-                    useNftDetection: false,
-                    openSeaEnabled: false,
-                    disabledRpcMethodPreferences: {
-                      eth_sign: false
-                    },
-                    _U: 0,
-                    _V: 1,
-                    _W: {
-                      featureFlags: {},
-                      frequentRpcList: [],
-                      identities: {
-                        '0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3': {
-                          address: "0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3",
-                          name: "Account 1",
-                          importTime: 1684232000456
-                        }
-                      },
-                      ipfsGateway: "https://cloudflare-ipfs.com/ipfs/",
-                      lostIdentities: {},
-                      selectedAddress: "0x76cf1CdD1fcC252442b50D6e97207228aA4aefC3",
-                      useTokenDetection: true,
-                      useNftDetection: false,
-                      openSeaEnabled: false,
-                      disabledRpcMethodPreferences: {
-                        eth_sign: false
-                      }
-                    },
-                    _X: null
-                  },
-                  TokenBalancesController: {
-                    contractBalances: {}
-                  },
-                  TokenRatesController: {
-                    contractExchangeRates: {}
-                  },
-                  TokensController: {
-                    tokens: [],
-                    ignoredTokens: [],
-                    detectedTokens: [],
-                    allTokens: {},
-                    allIgnoredTokens: {},
-                    allDetectedTokens: {},
-                    suggestedAssets: []
-                  },
-                  TransactionController: {
-                    methodData: {},
-                    transactions: [],
-                    internalTransactions: [],
-                    swapsTransactions: {}
-                  },
-                  TypedMessageManager: {
-                    unapprovedMessages: {},
-                    unapprovedMessagesCount: 0
-                  },
-                  SwapsController: {
-                    quotes: {},
-                    quoteValues: {},
-                    fetchParams: {
-                      slippage: 0,
-                      sourceToken: "",
-                      sourceAmount: 0,
-                      destinationToken: "",
-                      walletAddress: ""
-                    },
-                    fetchParamsMetaData: {
-                      sourceTokenInfo: {
-                        decimals: 0,
-                        address: "",
-                        symbol: ""
-                      },
-                      destinationTokenInfo: {
-                        decimals: 0,
-                        address: "",
-                        symbol: ""
-                      }
-                    },
-                    topAggSavings: null,
-                    aggregatorMetadata: null,
-                    tokens: null,
-                    topAssets: null,
-                    approvalTransaction: null,
-                    aggregatorMetadataLastFetched: 0,
-                    quotesLastFetched: 0,
-                    topAssetsLastFetched: 0,
-                    error: {
-                      key: null,
-                      description: null
-                    },
-                    topAggId: null,
-                    tokensLastFetched: 0,
-                    isInPolling: false,
-                    pollingCyclesLeft: 3,
-                    quoteRefreshSeconds: null,
-                    usedGasEstimate: null,
-                    usedCustomGas: null,
-                    chainCache: {
-                      1: {
-                        aggregatorMetadata: null,
-                        tokens: null,
-                        topAssets: null,
-                        aggregatorMetadataLastFetched: 0,
-                        topAssetsLastFetched: 0,
-                        tokensLastFetched: 0
-                      }
-                    }
-                  },
-                  GasFeeController: {
-                    gasFeeEstimates: {},
-                    estimatedGasFeeTimeBounds: {},
-                    gasEstimateType: "none"
-                  },
-                  TokenDetectionController: {},
-                  NftDetectionController: {},
-                  PermissionController: {
-                    subjects: {}
-                  }
-                }
-              },
-              privacy: {
-                approvedHosts: {},
-                thirdPartyApiMode: true,
-                revealSRPTimestamps: []
-              },
-              bookmarks: [],
-              recents: [],
-              browser: {
-                history: [],
-                whitelist: [],
-                tabs: [],
-                activeTab: null
-              },
-              modals: {
-                networkModalVisible: false,
-                shouldNetworkSwitchPopToWallet: true,
-                collectibleContractModalVisible: false,
-                receiveModalVisible: false,
-                dappTransactionModalVisible: false,
-                approveModalVisible: false
-              },
-              settings: {
-                searchEngine: "DuckDuckGo",
-                primaryCurrency: "ETH",
-                lockTime: 30000,
-                useBlockieIcon: true,
-                hideZeroBalanceTokens: false
-              },
-              alert: {
-                isVisible: false,
-                autodismiss: null,
-                content: null,
-                data: null
-              },
-              transaction: {
-                selectedAsset: {},
-                transaction: {}
-              },
-              user: {
-                loadingMsg: "",
-                loadingSet: false,
-                passwordSet: true,
-                seedphraseBackedUp: true,
-                backUpSeedphraseVisible: false,
-                protectWalletModalVisible: false,
-                gasEducationCarouselSeen: false,
-                nftDetectionDismissed: false,
-                userLoggedIn: true,
-                isAuthChecked: false,
-                initialScreen: "",
-                appTheme: "os"
-              },
-              wizard: {
-                step: 0
-              },
-              onboarding: {
-                events: []
-              },
-              notification: {
-                notifications: []
-              },
-              swaps: {
-                1: {
-                  isLive: true
-                },
-                isLive: true,
-                hasOnboarded: false
-              },
-              fiatOrders: {
-                orders: [],
-                customOrderIds: [],
-                networks: [
-                  {
-                    active: true,
-                    chainId: 1,
-                    chainName: "Ethereum Mainnet",
-                    shortName: "Ethereum",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 10,
-                    chainName: "Optimism Mainnet",
-                    shortName: "Optimism",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 25,
-                    chainName: "Cronos Mainnet",
-                    shortName: "Cronos",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 56,
-                    chainName: "BNB Chain Mainnet",
-                    shortName: "BNB Chain",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 137,
-                    chainName: "Polygon Mainnet",
-                    shortName: "Polygon",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 250,
-                    chainName: "Fantom Mainnet",
-                    shortName: "Fantom",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 1284,
-                    chainName: "Moonbeam Mainnet",
-                    shortName: "Moonbeam",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 1285,
-                    chainName: "Moonriver Mainnet",
-                    shortName: "Moonriver",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 42161,
-                    chainName: "Arbitrum Mainnet",
-                    shortName: "Arbitrum",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 42220,
-                    chainName: "Celo Mainnet",
-                    shortName: "Celo",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 43114,
-                    chainName: "Avalanche C-Chain Mainnet",
-                    shortName: "Avalanche C-Chain",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 1313161554,
-                    chainName: "Aurora Mainnet",
-                    shortName: "Aurora",
-                    nativeTokenSupported: false
-                  },
-                  {
-                    active: true,
-                    chainId: 1666600000,
-                    chainName: "Harmony Mainnet (Shard 0)",
-                    shortName: "Harmony  (Shard 0)",
-                    nativeTokenSupported: true
-                  },
-                  {
-                    active: true,
-                    chainId: 11297108109,
-                    chainName: "Palm Mainnet",
-                    shortName: "Palm",
-                    nativeTokenSupported: false
-                  }
-                ],
-                selectedRegionAgg: null,
-                selectedPaymentMethodAgg: null,
-                getStartedAgg: false,
-                authenticationUrls: [],
-                activationKeys: []
-              },
-              infuraAvailability: {
-                isBlocked: false
-              },
-              navigation: {
-                currentRoute: "AdvancedSettings",
-                currentBottomNavRoute: "Wallet"
-              },
-              networkOnboarded: {
-                networkOnboardedState: [],
-                networkState: {
-                  showNetworkOnboarding: false,
-                  nativeToken: "",
-                  networkType: "",
-                  networkUrl: ""
-                },
-                switchedNetwork: {
-                  networkUrl: "",
-                  networkStatus: false
-                }
-              },
-              security: {
-                allowLoginWithRememberMe: false,
-                automaticSecurityChecksEnabled: false,
-                hasUserSelectedAutomaticSecurityCheckOption: true,
-                isAutomaticSecurityChecksModalOpen: false
-              }
-          },
-          asyncState: {
-            '@MetaMask:existingUser': "true",
-            '@MetaMask:onboardingWizard': "explored",
-            '@MetaMask:UserTermsAcceptedv1.0': "true",
-            '@MetaMask:WhatsNewAppVersionSeen': "6.5.0"
-          }
-      };
-      return this;
-    }
-  
-    /**
-     * Set the fixture to an empty object for onboarding.
-     * @returns {FixtureBuilder} - The FixtureBuilder instance for method chaining.
-     */
-    withOnboardingFixture() {
-      this.fixture = {
-        asyncState:{}
-      };
-      return this;
-    }
-  
-    /**
-     * Build and return the fixture object.
-     * @returns {Object} - The built fixture object.
-     */
-    build() {
-      return this.fixture;
-    }
-  }
-
-  module.exports = FixtureBuilder;
